e2e: serial: split resourcelist package
move resourcelist utilities in their own helper.
