+++ test/e2e/serial/workload_placement_test.go
@@ -45,6 +45,7 @@ import (
 	"github.com/openshift-kni/numaresources-operator/test/utils/objects"
 	e2ewait "github.com/openshift-kni/numaresources-operator/test/utils/objects/wait"
 	e2epadder "github.com/openshift-kni/numaresources-operator/test/utils/padder"
+	e2ereslist "github.com/openshift-kni/numaresources-operator/test/utils/resourcelist"
 )
 
 const testKey = "testkey"
@@ -100,7 +101,7 @@ var _ = Describe("[serial][disruptive][scheduler] workload placement", func() {
 				corev1.ResourceMemory: resource.MustParse("8Gi"),
 			}
 
-			By(fmt.Sprintf("creating test pod, total resources required %s", e2enrt.ResourceListToString(requiredRes)))
+			By(fmt.Sprintf("creating test pod, total resources required %s", e2ereslist.ToString(requiredRes)))
 
 			// TODO: we need AT LEAST 2 (so 4, 8 is fine...) but we hardcode the padding logic to keep the test simple,
 			// so we can't support ATM zones > 2. HW with zones > 2 is rare anyway, so not to big of a deal now.
@@ -212,7 +213,7 @@ var _ = Describe("[serial][disruptive][scheduler] workload placement", func() {
 			time.Sleep(18 * time.Second)
 			dumpNRTForNode(fxt.Client, targetNodeName)
 
-			By(fmt.Sprintf("running the test pod requiring: %s", e2enrt.ResourceListToString(requiredRes)))
+			By(fmt.Sprintf("running the test pod requiring: %s", e2ereslist.ToString(requiredRes)))
 			pod := objects.NewTestPodPause(fxt.Namespace.Name, "testpod")
 			pod.Spec.SchedulerName = schedulerName
 			pod.Spec.Containers[0].Resources.Limits = requiredRes
@@ -276,7 +277,7 @@ var _ = Describe("[serial][disruptive][scheduler] workload placement", func() {
 				corev1.ResourceCPU:    resource.MustParse("12"),
 				corev1.ResourceMemory: resource.MustParse("8Gi"),
 			}
-			requiredRes := e2enrt.ResourcesFromGuaranteedPod(*pod)
+			requiredRes := e2ereslist.FromGuaranteedPod(*pod)
 
 			// make sure the sum is equal to the sum of the requirement of the test pod,
 			// so the *node* total free resources are equal between the target node and
@@ -326,7 +327,7 @@ var _ = Describe("[serial][disruptive][scheduler] workload placement", func() {
 				zone := nrtInfo.Zones[idx]
 				cnt := pod.Spec.Containers[1-idx] // switch requirements intentionally - both couplings are legit anyway
 
-				By(fmt.Sprintf("padding node %q zone %q to fit only %s", nrtInfo.Name, zone.Name, e2enrt.ResourceListToString(cnt.Resources.Limits)))
+				By(fmt.Sprintf("padding node %q zone %q to fit only %s", nrtInfo.Name, zone.Name, e2ereslist.ToString(cnt.Resources.Limits)))
 				padPod, err := makePaddingPod(fxt.Namespace.Name, "target", zone, cnt.Resources.Limits)
 				Expect(err).ToNot(HaveOccurred())
 
@@ -347,7 +348,7 @@ var _ = Describe("[serial][disruptive][scheduler] workload placement", func() {
 					padRes := unsuitableFreeRes[zoneIdx]
 
 					name := fmt.Sprintf("unsuitable%d", nodeIdx)
-					By(fmt.Sprintf("saturating node %q -> %q zone %q to fit only %s", nrtInfo.Name, name, zone.Name, e2enrt.ResourceListToString(padRes)))
+					By(fmt.Sprintf("saturating node %q -> %q zone %q to fit only %s", nrtInfo.Name, name, zone.Name, e2ereslist.ToString(padRes)))
 					padPod, err := makePaddingPod(fxt.Namespace.Name, name, zone, padRes)
 					Expect(err).ToNot(HaveOccurred())
 
@@ -680,7 +681,7 @@ var _ = Describe("[serial][disruptive][scheduler] workload placement", func() {
 			nrtPostCreateList, err := e2enrt.GetUpdated(fxt.Client, nrtInitialList, time.Second*10)
 			Expect(err).ToNot(HaveOccurred())
 
-			rl := e2enrt.ResourcesFromGuaranteedPod(*updatedPod)
+			rl := e2ereslist.FromGuaranteedPod(*updatedPod)
 
 			nrtInitial, err := e2enrt.FindFromList(nrtInitialList.Items, updatedPod.Spec.NodeName)
 			Expect(err).ToNot(HaveOccurred())
@@ -841,7 +842,7 @@ var _ = Describe("[serial][disruptive][scheduler] workload placement", func() {
 			Expect(err).ToNot(HaveOccurred())
 			Expect(schedOK).To(BeTrue(), "pod %s/%s not scheduled with expected scheduler %s", updatedPod.Namespace, updatedPod.Name, schedulerName)
 
-			rl := e2enrt.ResourcesFromGuaranteedPod(updatedPod)
+			rl := e2ereslist.FromGuaranteedPod(updatedPod)
 
 			nrtInitial, err := e2enrt.FindFromList(nrtInitialList.Items, updatedPod.Spec.NodeName)
 			Expect(err).ToNot(HaveOccurred())
@@ -901,7 +902,7 @@ var _ = Describe("[serial][disruptive][scheduler] workload placement", func() {
 			Expect(err).ToNot(HaveOccurred())
 			Expect(schedOK).To(BeTrue(), "pod %s/%s not scheduled with expected scheduler %s", updatedPod.Namespace, updatedPod.Name, schedulerName)
 
-			rl = e2enrt.ResourcesFromGuaranteedPod(updatedPod)
+			rl = e2ereslist.FromGuaranteedPod(updatedPod)
 
 			nrtPostCreate, err = e2enrt.FindFromList(nrtPostCreateDeploymentList.Items, updatedPod.Spec.NodeName)
 			Expect(err).ToNot(HaveOccurred())
@@ -991,7 +992,7 @@ var _ = Describe("[serial][disruptive][scheduler] workload placement", func() {
 			Expect(err).ToNot(HaveOccurred())
 			Expect(schedOK).To(BeTrue(), "pod %s/%s not scheduled with expected scheduler %s", updatedPod.Namespace, updatedPod.Name, schedulerName)
 
-			rl = e2enrt.ResourcesFromGuaranteedPod(updatedPod)
+			rl = e2ereslist.FromGuaranteedPod(updatedPod)
 
 			nrtLastUpdate, err := e2enrt.FindFromList(nrtLastUpdateDeploymentList.Items, updatedPod.Spec.NodeName)
 			Expect(err).ToNot(HaveOccurred())
@@ -1028,14 +1029,14 @@ var _ = Describe("[serial][disruptive][scheduler] workload placement", func() {
 })
 
 func makePaddingPod(namespace, nodeName string, zone nrtv1alpha1.Zone, podReqs corev1.ResourceList) (*corev1.Pod, error) {
-	klog.Infof("want to have zone %q with allocatable: %s", zone.Name, e2enrt.ResourceListToString(podReqs))
+	klog.Infof("want to have zone %q with allocatable: %s", zone.Name, e2ereslist.ToString(podReqs))
 
 	paddingReqs, err := e2enrt.SaturateZoneUntilLeft(zone, podReqs)
 	if err != nil {
 		return nil, err
 	}
 
-	klog.Infof("padding resource to saturate %q: %s", nodeName, e2enrt.ResourceListToString(paddingReqs))
+	klog.Infof("padding resource to saturate %q: %s", nodeName, e2ereslist.ToString(paddingReqs))
 
 	padPod := objects.NewTestPodPause(namespace, fmt.Sprintf("padpod-%s-%s", nodeName, zone.Name))
 	padPod.Spec.Containers[0].Resources.Limits = paddingReqs
